UnifiedWorkflow/app/sequentialthinking_service/requirements.txt
# Sequential Thinking Service Dependencies
# FastAPI and async framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# LangGraph and state management
langgraph==0.0.69
langchain==0.1.0
langchain-core==0.1.0
langchain-community==0.0.10

# Redis for checkpointing
redis==5.0.1
redis[hiredis]==5.0.1

# HTTP client and async support
aiohttp==3.9.1
httpx==0.25.2

# Authentication
pyjwt==2.8.0
cryptography==41.0.8

# Data processing and validation
numpy==1.24.4
pandas==2.0.3

# Utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0

# Logging and monitoring
structlog==23.2.0
python-json-logger==2.0.7

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Memory optimization
objsize==0.6.1

# Additional reasoning tools
typing-extensions==4.8.0
uuid==1.30